vite-project/backend/requirements.txt
pandas
keybert
numpy
fuzzywuzzy
python-Levenshtein  # recommended for better fuzzywuzzy performance
flask
flask-cors
openai
python-dotenv